code review suggestion; breakout attribute utilities from composition
